Update ncd_enterprise.py
The data was starting at the wrong index. This has been fixed. ODR is now passed as well as the NodeId.

The if statement dictating what should be added to the read buffer was improperly formatted. It has been improved to break off computation when the end of the packet has been reached.